refactor(pytest): establish green collection baseline after jvi pydantic migration
jvi migrated its geo types (Size/Rect/Point) and label models to pydantic
with keyword-only init, and tightened frozen dataclass contracts. This broke
jxl's pytest collection (14 errors) and several stale runtime tests.

Collection fixes (14 -> 0 errors):
- hop.py LabelRecord(FileRecord): frozen=True (FileRecord now frozen in jvi;
  fixes 'cannot inherit non-frozen dataclass from a frozen one')
- positional Size(...)/Rect(...) constructors -> Size.new(...)/Rect.new(...)
  across diag_extractor, cutter2, dataset_checker, blend, tile/blend tests
- pyproject [tool.pytest.ini_options]: ignore 10 dead test modules that
  import physically-removed deps (pandas/autogluon/torchsummary/ias/jcx.m/
  jxl.det.idetector/yolo_old) - documented, non-silent (no fallback path)

Runtime fixes:
- classifier.vote_weighted: cast probs ndarray to float for robust
  normalization regardless of input dtype (behavior-preserving)
- d2d_test: add required id field (jvi D2dObject contract)

6 stale tests skipped with FIXME (jvi-migration API drift / external data):
  d2d_test (D2dObjectTrack removed), diag_extractor_test (signature drift +
  external data), info_test (A2dImageLabel.new removed), hop_test (external
  data dirs), meta_test (META_DIR non-LabelMeta json). Defer to PR-3.

Result: 17 passed, 6 skipped, 0 collection errors.

+++ src/jxl/cls/classifier.py
@@ -80,7 +80,7 @@ class ClassifierRes(Protocol):
 
 def vote_weighted(arr: list[ClassifierRes]) -> ClassifierRes:
     """根据概率投票获取最终分类结果"""
-    mat = np.array([a.confidences() for a in arr])
+    mat = np.array([a.confidences() for a in arr], dtype=float)
     v = np.sum(mat, axis=0)
     v /= np.sum(v)
     return ClassifierResList(probs=v.tolist())
